requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
requests==2.31.0
apscheduler==3.10.4
fastapi-cors==0.0.6